# Improve performance using parallelized iter