--- a/native/bls/Cargo.toml
+++ b/native/bls/Cargo.toml
@@ -18,3 +18,4 @@ bls12_381 = { version = "0.8.0", features = ["experimental"] }
 pairing = "0.23.0"
 group = "0.13.0"
 sha2 = "0.10.8"
+rayon = "1.10.0"